requirements.txt
annotated-types==0.7.0
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.30
sqlmodel==0.0.19
typing_extensions==4.12.2
Werkzeug==3.0.3
